Fully integrated drop down for major city selection over home and forecast detail pages
