+++ app/src/main/res/layout/activity_forecast_detail.xml
@@ -11,16 +11,18 @@
         android:layout_width="0dp"
         android:layout_height="wrap_content"
         android:layout_marginStart="8dp"
-        android:layout_marginTop="32dp"
+        android:layout_marginTop="40dp"
         android:layout_marginEnd="8dp"
         android:autoLink="none"
         android:autoSizeMaxTextSize="30sp"
         android:autoSizeMinTextSize="10sp"
         android:autoSizeStepGranularity="1sp"
         android:autoSizeTextType="uniform"
-        android:digits="abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ "
         android:fontFamily="sans-serif-black"
-        android:inputType="textNoSuggestions"
+        android:focusable="false"
+        android:longClickable="false"
+        android:clickable="false"
+        android:inputType="none"
         android:maxLines="1"
         android:padding="8dp"
         android:textAlignment="center"
@@ -38,7 +40,7 @@
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:layout_marginStart="128dp"
-        android:layout_marginTop="32dp"
+        android:layout_marginTop="40dp"
         android:layout_marginEnd="128dp"
         android:fontFamily="sans-serif-black"
         android:textAlignment="center"
@@ -53,7 +55,7 @@
         android:id="@+id/feels_like_text_view"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:layout_marginTop="32dp"
+        android:layout_marginTop="40dp"
         android:fontFamily="sans-serif-medium"
         android:textAlignment="center"
         android:textSize="24sp"
